Use precise line numbers in trace test. (#3023)
Don't replace line numbers with offsets from the CHECK line, because
that's both hard to read and unstable when the trace output changes.
Instead, include the actual line numbers. Move all the CHECKs to the end
of the file so that they don't reference lines that appear after them.

This is intended to reduce the number of merge conflicts in
full_trace.carbon and make the test CHECKs more readable.

---------

Co-authored-by: Jon Ross-Perkins <jperkins@google.com>