Fix spelling error in README (#3069)
This PR resolves the spelling error in the "Contributing" section of the
README. The sentence "Contribute the language design: feedback on
design, new design proposal" has been corrected to "Contribute to the
language design: feedback on design, new design proposal". This change
enhances the clarity and accuracy of the contribution guidelines.

Closes #3068

Co-authored-by: aroramrinaal <aroramrinaal@gmail.com>