Improve diagnosis of mismatched brackets.
Move handling of mismatched brackets out of the main lexing loop
into a separate pass that is only run if there are mismatched brackets,
in preparation for using both lookahead and lookbehind to work out how
to match brackets.

Fix bracket insertion location to be immediately after the token that
we're inserting the bracket after, rather than potentially at the end of
a comment.

When there are open brackets at the end of the file, say that there are
open brackets, not that there's a closing bracket without a matching
opening bracket.