Introduce `bazel-diff` tool and infra.
This adds access to the `bazel-diff` tool that can be used to compute
hashes of all inputs to the Bazel build, and then diff them to compute
the possibly impacted set of targets.

This then adds a Python script that uses this to compute the impacted
target set between two Git commits.

And finally, it adds a stage in our CI to try to compute this impacted
target set.

Currently, this isn't used as I want to be able to see how it works in
practice and whether it computes the correct difference. Also, I'm
interested in when and where folks would like to use something like
this.

Once we get it working, this would let us run *much* more efficient
build and test actions in CI. The way `bazel-diff` works is that it just
queries Bazel for the input hashes without running any actions. This is
very fast and local. We have very good cache hits in our CI already for
unchanged parts of the tree, but the actions still have to download all
of those cached inputs. As the project is growing, this gets *extremely*
large (several gigabytes). With this, we only need to fetch the inputs
necessary to build and run any potentially impacted targets.