Update C++Now talks with video links (#3158)
The successor strategy talk isn't up yet, but will be on Sep 8 -- just
trying to get them all in one pass.

Slides were previously linked, which I could keep, although it's weird
to have a long link with just `(Slides)` at the end because it implies
the whole linked text is slides (rather than a video). So if that's
desired, I might switch instead to something like:

```
- Talk name, con ([video](youtube link), [slides](slides link))
```

But, this current format (video-only full-name link) felt consistent
with the 2022 links.